Re #893 fixed test_sqw_file tests for current modifications

+++ _test/test_sqw_file/test_binfile_v2_common.m
@@ -136,7 +136,7 @@ classdef test_binfile_v2_common <  TestCase %WithSave
             tob=tob.set_file_to_update(samp);
 
             assertTrue(tob.sqw_type)
-            assertEqual(tob.num_dim,int32(1))
+            assertEqual(tob.num_dim,1)
             assertTrue(isa(tob,'faccess_sqw_v2'));
 
             tob = tob.delete();
